requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
python-decouple==3.8
Pillow==10.1.0
reportlab==4.0.4
geopy==2.4.0 